# Feature #17: CLI Interface & User Experience - Complete Implementation