… coverage/coverage-final.json
… coverage/index.html
… coverage/lcov-report/index.html
… coverage/lcov-report/src/achievement-system.js.html
… coverage/lcov-report/src/character/index.html
… coverage/lcov-report/src/character/ravi.js.html
… coverage/lcov-report/src/command-parser.js.html
… coverage/lcov-report/src/dialogue-system.js.html
… coverage/lcov-report/src/game-engine.js.html
… coverage/lcov-report/src/game-state.js.html
… coverage/lcov-report/src/gameplay-integration.js.html
… coverage/lcov-report/src/index.html
… coverage/lcov-report/src/narrative-controller.js.html
… coverage/lcov-report/src/puzzle-system.js.html
… coverage/lcov-report/src/stories/index.html
… coverage/lcov-report/src/stories/index.js.html
… coverage/lcov-report/src/story-engine.js.html
… coverage/lcov.info
… coverage/src/achievement-system.js.html
… coverage/src/character/index.html
… coverage/src/character/ravi.js.html
… coverage/src/command-parser.js.html
… coverage/src/dialogue-system.js.html
… coverage/src/game-engine.js.html
… coverage/src/game-state.js.html
… coverage/src/gameplay-integration.js.html
… coverage/src/index.html
… coverage/src/narrative-controller.js.html
… coverage/src/puzzle-system.js.html
… coverage/src/stories/index.html