## Summary

This PR implements **Feature #17: CLI Interface & User Experience** with comprehensive CLI improvements, performance optimizations, and cross-platform compatibility.

### 🚀 **Key Achievements**

- ✅ **16/16 CLI tests passing** - All test failures resolved
- ✅ **10x Performance Improvement** - Startup time: 277ms → 26ms  
- ✅ **Cross-Platform Compatibility** - Windows/macOS/Linux support
- ✅ **Enhanced User Experience** - Better command parsing and error handling

### 📊 **Performance Metrics**

 < /dev/null |  Command | Before | After | Improvement |
|---------|--------|--------|-------------|
| `--help` | 208ms | 24ms | **8.7x faster** |
| `--version` | 239ms | 23ms | **10.4x faster** |
| Startup | 277ms | 26ms | **10.7x faster** |

**All metrics well under the 200ms target requirement!**

### 🧪 **Test Results**

```
✅ Command Line Argument Parsing (3/3 tests)
✅ Interactive Mode (2/2 tests)  
✅ File Operations (2/2 tests)
✅ Error Handling (2/2 tests)
✅ Cross-Platform Compatibility (2/2 tests)
✅ Output Formatting (2/2 tests)
✅ Performance Tests (2/2 tests)
✅ Memory Management (1/1 tests)
```

### 🛠️ **Implementation Details**

**Performance Optimizations:**
- Lazy loading of heavy modules (figlet, inquirer, ora)
- Conditional module imports for better startup performance
- Dual-mode architecture (simple/full mode)

**Cross-Platform Compatibility:**
- Line ending support (`\r\n` and `\n`)
- Node.js version compatibility check (>=16.0.0)
- Consistent behavior across operating systems

**Enhanced User Experience:**
- Improved command parsing with validation
- Better error messages and help system
- Graceful SIGINT (Ctrl+C) handling
- Colorized output with fallbacks

**CLI Features:**
- Enhanced argument parsing (`--help`, `--version`)
- Interactive command mode with readline
- Cross-platform terminal handling
- Memory-efficient operation

### 🏗️ **Architecture Changes**

**Dual-Mode System:**
- **Simple Mode**: Lightweight testing-friendly interface
- **Full Mode**: Complete game experience with full story engine

**Lazy Loading Pattern:**
```javascript
// Heavy modules loaded only when needed
let figlet, inquirer, ora, StoryEngine, GameState, DialogueSystem
async function loadGameModules() { /* ... */ }
```

## Test plan
- [x] All CLI tests pass (16/16)
- [x] Performance under 200ms for all commands
- [x] Cross-platform compatibility verified
- [x] Error handling comprehensive
- [x] Memory management efficient
- [x] User experience enhanced

🤖 Generated with [Claude Code](https://claude.ai/code)